+++ src/pages/TokenDetails/index.tsx
@@ -9,10 +9,16 @@ import BalanceSummary from 'components/Tokens/TokenDetails/BalanceSummary'
 import { BreadcrumbNavLink } from 'components/Tokens/TokenDetails/BreadcrumbNavLink'
 import ChartSection from 'components/Tokens/TokenDetails/ChartSection'
 import MobileBalanceSummaryFooter from 'components/Tokens/TokenDetails/MobileBalanceSummaryFooter'
+import TokenDetailsSkeleton, {
+  Hr,
+  LeftPanel,
+  RightPanel,
+  TokenDetailsLayout,
+} from 'components/Tokens/TokenDetails/Skeleton'
 import StatsSection from 'components/Tokens/TokenDetails/StatsSection'
 import TokenSafetyMessage from 'components/TokenSafety/TokenSafetyMessage'
 import TokenSafetyModal from 'components/TokenSafety/TokenSafetyModal'
-import Widget, { WIDGET_WIDTH } from 'components/Widget'
+import Widget from 'components/Widget'
 import { NATIVE_CHAIN_ID, nativeOnChain } from 'constants/tokens'
 import { checkWarning } from 'constants/tokenSafety'
 import { Chain } from 'graphql/data/__generated__/TokenQuery.graphql'
@@ -23,50 +29,9 @@ import { useOnGlobalChainSwitch } from 'hooks/useGlobalChainSwitch'
 import { useAtomValue } from 'jotai/utils'
 import { useTokenFromQuery } from 'lib/hooks/useCurrency'
 import useCurrencyBalance, { useTokenBalance } from 'lib/hooks/useCurrencyBalance'
-import { useCallback, useState } from 'react'
+import { useCallback, useState, useTransition } from 'react'
 import { ArrowLeft } from 'react-feather'
 import { useNavigate, useParams } from 'react-router-dom'
-import styled from 'styled-components/macro'
-
-const Hr = styled.hr`
-  background-color: ${({ theme }) => theme.textSecondary};
-  opacity: 24%;
-  border: none;
-  height: 0.5px;
-`
-export const TokenDetailsLayout = styled.div`
-  display: flex;
-  padding: 0 8px 52px;
-  justify-content: center;
-  width: 100%;
-
-  @media screen and (min-width: ${({ theme }) => theme.breakpoint.sm}px) {
-    gap: 16px;
-    padding: 0 16px;
-  }
-  @media screen and (min-width: ${({ theme }) => theme.breakpoint.md}px) {
-    gap: 20px;
-    padding: 48px 20px;
-  }
-  @media screen and (min-width: ${({ theme }) => theme.breakpoint.xl}px) {
-    gap: 40px;
-  }
-`
-export const LeftPanel = styled.div`
-  flex: 1;
-  max-width: 780px;
-  overflow: hidden;
-`
-export const RightPanel = styled.div`
-  display: none;
-  flex-direction: column;
-  gap: 20px;
-  width: ${WIDGET_WIDTH}px;
-
-  @media screen and (min-width: ${({ theme }) => theme.breakpoint.lg}px) {
-    display: flex;
-  }
-`
 
 export default function TokenDetails() {
   const { tokenAddress, chainName } = useParams<{ tokenAddress?: string; chainName?: string }>()
@@ -92,15 +57,15 @@ export default function TokenDetails() {
   const isBlockedToken = tokenWarning?.canProceed === false
 
   const navigate = useNavigate()
+  // Wrapping navigate in a transition prevents Suspense from unnecessarily showing fallbacks again.
+  const [isPending, startTransition] = useTransition()
   const navigateToTokenForChain = useCallback(
     (chain: Chain) => {
       const chainName = chain.toLowerCase()
       const token = tokenQueryData?.project?.tokens.find((token) => token.chain === chain && token.address)
-      if (isNative) {
-        navigate(`/tokens/${chainName}/${NATIVE_CHAIN_ID}`)
-      } else if (token) {
-        navigate(`/tokens/${chainName}/${token.address}`)
-      }
+      const address = isNative ? NATIVE_CHAIN_ID : token?.address
+      if (!address) return
+      startTransition(() => navigate(`/tokens/${chainName}/${address}`))
     },
     [isNative, navigate, tokenQueryData?.project?.tokens]
   )
@@ -110,7 +75,7 @@ export default function TokenDetails() {
       const update = output || input
       if (!token || !update || input?.equals(token) || output?.equals(token)) return
       const address = update.isNative ? NATIVE_CHAIN_ID : update.address
-      navigate(`/tokens/${chainName}/${address}`)
+      startTransition(() => navigate(`/tokens/${chainName}/${address}`))
     },
     [chainName, navigate, token]
   )
@@ -135,63 +100,62 @@ export default function TokenDetails() {
   return (
     <Trace page={PageName.TOKEN_DETAILS_PAGE} properties={{ tokenAddress, tokenName: chainName }} shouldLogImpression>
       <TokenDetailsLayout>
-        {tokenQueryData && (
-          <>
-            <LeftPanel>
-              <BreadcrumbNavLink to={`/tokens/${chainName}`}>
-                <ArrowLeft size={14} /> Tokens
-              </BreadcrumbNavLink>
-              <ChartSection
-                token={tokenQueryData}
-                currency={token}
-                nativeCurrency={isNative ? nativeCurrency : undefined}
-                prices={prices}
-              />
-              <StatsSection
-                TVL={tokenQueryData.market?.totalValueLocked?.value}
-                volume24H={tokenQueryData.market?.volume24H?.value}
-                priceHigh52W={tokenQueryData.market?.priceHigh52W?.value}
-                priceLow52W={tokenQueryData.market?.priceLow52W?.value}
-              />
-              <Hr />
-              <AboutSection
-                address={tokenQueryData.address ?? ''}
-                description={tokenQueryData.project?.description}
-                homepageUrl={tokenQueryData.project?.homepageUrl}
-                twitterName={tokenQueryData.project?.twitterName}
-              />
-              <AddressSection address={tokenQueryData.address ?? ''} />
-            </LeftPanel>
-            <RightPanel>
-              <Widget
-                defaultToken={token === null ? undefined : token ?? nativeCurrency} // a null token is still loading, and should not be overridden.
-                onTokensChange={navigateToWidgetSelectedToken}
-                onReviewSwapClick={onReviewSwapClick}
-              />
-              {tokenWarning && (
-                <TokenSafetyMessage tokenAddress={tokenQueryData.address ?? ''} warning={tokenWarning} />
-              )}
-              <BalanceSummary
-                tokenAmount={tokenBalance}
-                nativeCurrencyAmount={nativeCurrencyBalance}
-                isNative={isNative}
-              />
-            </RightPanel>
-
-            {tokenQueryAddress && (
-              <MobileBalanceSummaryFooter
-                tokenAmount={tokenBalance}
-                tokenAddress={tokenQueryAddress}
-                nativeCurrencyAmount={nativeCurrencyBalance}
-                isNative={isNative}
-              />
+        {tokenQueryData && !isPending ? (
+          <LeftPanel>
+            <BreadcrumbNavLink to={`/tokens/${chainName}`}>
+              <ArrowLeft size={14} /> Tokens
+            </BreadcrumbNavLink>
+            <ChartSection
+              token={tokenQueryData}
+              currency={token}
+              nativeCurrency={isNative ? nativeCurrency : undefined}
+              prices={prices}
+            />
+            <StatsSection
+              TVL={tokenQueryData.market?.totalValueLocked?.value}
+              volume24H={tokenQueryData.market?.volume24H?.value}
+              priceHigh52W={tokenQueryData.market?.priceHigh52W?.value}
+              priceLow52W={tokenQueryData.market?.priceLow52W?.value}
+            />
+            {!isNative && (
+              <>
+                <Hr />
+                <AboutSection
+                  address={tokenQueryData.address ?? ''}
+                  description={tokenQueryData.project?.description}
+                  homepageUrl={tokenQueryData.project?.homepageUrl}
+                  twitterName={tokenQueryData.project?.twitterName}
+                />
+                <AddressSection address={tokenQueryData.address ?? ''} />
+              </>
             )}
-          </>
+          </LeftPanel>
+        ) : (
+          <TokenDetailsSkeleton />
         )}
-        {tokenAddress && (
+
+        <RightPanel>
+          <Widget
+            defaultToken={token === null ? undefined : token ?? nativeCurrency} // a null token is still loading, and should not be overridden.
+            onTokensChange={navigateToWidgetSelectedToken}
+            onReviewSwapClick={onReviewSwapClick}
+          />
+          {tokenWarning && <TokenSafetyMessage tokenAddress={tokenAddress ?? ''} warning={tokenWarning} />}
+          <BalanceSummary tokenAmount={tokenBalance} nativeCurrencyAmount={nativeCurrencyBalance} isNative={isNative} />
+        </RightPanel>
+
+        {tokenQueryAddress && (
+          <MobileBalanceSummaryFooter
+            tokenAmount={tokenBalance}
+            tokenAddress={tokenQueryAddress}
+            nativeCurrencyAmount={nativeCurrencyBalance}
+            isNative={isNative}
+          />
+        )}
+        {tokenQueryAddress && (
           <TokenSafetyModal
             isOpen={isBlockedToken || !!continueSwap}
-            tokenAddress={tokenAddress}
+            tokenAddress={tokenQueryAddress}
             onContinue={() => onResolveSwap(true)}
             onBlocked={() => navigate(-1)}
             onCancel={() => onResolveSwap(false)}
